Added a displayTable method
this method displays the data in a grid so when showing large amounts of data like with the inventory we can display it as a grid so the user can read the data more easily.
We do lose some customizability because we are abstracting away a lot.
Individual processes don't get to assume they are printing to the console so we can't use tabs and new line charaters for formatting